+++ MatrixSDK/Background/MXBackgroundSyncService.swift
@@ -81,20 +81,16 @@ public enum MXBackgroundSyncServiceError: Error {
         // We can flush any crypto data if our sync response store is empty
         let resetBackgroundCryptoStore = syncResponseStoreManager.syncToken() == nil
         
-        #if DEBUG
-        if MXSDKOptions.sharedInstance().enableCryptoV2 {
-            do {
-                crypto = try MXBackgroundCryptoV2(credentials: credentials, restClient: restClient)
-            } catch {
-                MXLog.failure("[MXBackgroundSyncService] init: Cannot initialize crypto v2", context: error)
-                crypto = MXLegacyBackgroundCrypto(credentials: credentials, resetBackgroundCryptoStore: resetBackgroundCryptoStore)
+        crypto = {
+            #if DEBUG
+            if MXSDKOptions.sharedInstance().isCryptoSDKAvailable && MXSDKOptions.sharedInstance().enableCryptoSDK {
+                // Crypto V2 is currently unable to decrypt notifications due to single-process store,
+                // so it uses dummy background crypto that does not do anything.
+                return MXDummyBackgroundCrypto()
             }
-        } else {
-            crypto = MXLegacyBackgroundCrypto(credentials: credentials, resetBackgroundCryptoStore: resetBackgroundCryptoStore)
-        }
-        #else
-        crypto = MXLegacyBackgroundCrypto(credentials: credentials, resetBackgroundCryptoStore: resetBackgroundCryptoStore)
-        #endif
+            #endif
+            return MXLegacyBackgroundCrypto(credentials: credentials, resetBackgroundCryptoStore: resetBackgroundCryptoStore)
+        }()
         
         pushRulesManager = MXBackgroundPushRulesManager(withCredentials: credentials)
         if let accountData = syncResponseStoreManager.syncResponseStore.accountData {
@@ -395,16 +391,18 @@ public enum MXBackgroundSyncServiceError: Error {
                     return
                 }
 
-                self.handleSyncResponse(syncResponse, syncToken: eventStreamToken)
-                
-                if let event = self.syncResponseStoreManager.event(withEventId: eventId, inRoom: roomId),
-                    !self.crypto.canDecryptEvent(event),
-                    (syncResponse.toDevice?.events ?? []).count > 0 {
-                    //  we got the event but not the keys to decrypt it. continue to sync
-                    self.launchBackgroundSync(forEventId: eventId, roomId: roomId, completion: completion)
-                } else {
-                    //  do not allow to sync anymore
-                    self._event(withEventId: eventId, inRoom: roomId, allowSync: false, completion: completion)
+                Task {
+                    await self.handleSyncResponse(syncResponse, syncToken: eventStreamToken)
+                    
+                    if let event = self.syncResponseStoreManager.event(withEventId: eventId, inRoom: roomId),
+                       !self.crypto.canDecryptEvent(event),
+                       (syncResponse.toDevice?.events ?? []).count > 0 {
+                        //  we got the event but not the keys to decrypt it. continue to sync
+                        self.launchBackgroundSync(forEventId: eventId, roomId: roomId, completion: completion)
+                    } else {
+                        //  do not allow to sync anymore
+                        self._event(withEventId: eventId, inRoom: roomId, allowSync: false, completion: completion)
+                    }
                 }
             case .failure(let error):
                 guard let _ = self else {
@@ -422,7 +420,7 @@ public enum MXBackgroundSyncServiceError: Error {
         }
     }
     
-    private func handleSyncResponse(_ syncResponse: MXSyncResponse, syncToken: String) {
+    private func handleSyncResponse(_ syncResponse: MXSyncResponse, syncToken: String) async {
         MXLog.debug("""
             [MXBackgroundSyncService] handleSyncResponse: \
             Received \(syncResponse.rooms?.join?.count ?? 0) joined rooms, \
@@ -436,7 +434,7 @@ public enum MXBackgroundSyncServiceError: Error {
         }
         syncResponseStoreManager.updateStore(with: syncResponse, syncToken: syncToken)
         
-        crypto.handleSyncResponse(syncResponse)
+        await crypto.handleSyncResponse(syncResponse)
         
         if MXSDKOptions.sharedInstance().autoAcceptRoomInvites,
            let invitedRooms = syncResponse.rooms?.invite {