+++ MatrixSDK/Data/MXRoom.m
@@ -2151,6 +2151,30 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
     return [mxSession.matrixRestClient redactEvent:eventId inRoom:self.roomId reason:reason success:success failure:failure];
 }
 
+- (MXHTTPOperation*)redactEvent:(NSString*)eventId
+                  withRelations:(NSArray<NSString *>*)relations
+                         reason:(NSString*)reason
+                        success:(void (^)(void))success
+                        failure:(void (^)(NSError *error))failure
+{
+    BOOL isFeatureWithRelationsSupported = (mxSession.store.supportedMatrixVersions.supportsRedactionWithRelations || mxSession.store.supportedMatrixVersions.supportsRedactionWithRelationsUnstable);
+    BOOL isFeatureWithRelationsStable = mxSession.store.supportedMatrixVersions.supportsRedactionWithRelations;
+
+    if ((relations != nil || [relations count] > 0) && !isFeatureWithRelationsSupported)
+    {
+        MXLogDebug(@"[MXRoom] redaction with relations is not supported (MSC3912)");
+    }
+    
+    return [mxSession.matrixRestClient redactEvent:eventId
+                                            inRoom:self.roomId
+                                            reason:reason
+                                             txnId:nil
+                                     withRelations:isFeatureWithRelationsSupported ? relations : nil
+                             withRelationsIsStable:isFeatureWithRelationsStable
+                                           success:success
+                                           failure:failure];
+}
+
 - (MXHTTPOperation *)reportEvent:(NSString *)eventId
                            score:(NSInteger)score
                           reason:(NSString *)reason
@@ -2271,6 +2295,12 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
         
         senderMessageBody = question;
     }
+    if (eventToReply.eventType == MXEventTypePollEnd)
+    {
+        // The "Ended poll" text is not meant to be localized from the sender side.
+        // This is why here we use a "default localizer" providing the english version of it.
+        senderMessageBody = MXSendReplyEventDefaultStringLocalizer.new.replyToEndedPoll;
+    }
     else if (eventToReply.eventType == MXEventTypeBeaconInfo)
     {
         senderMessageBody = stringLocalizer.senderSentTheirLiveLocation;
@@ -2345,8 +2375,7 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
         *replyContentFormattedBody = [self replyMessageFormattedBodyFromEventToReply:eventToReply
                                                           senderMessageFormattedBody:senderMessageFormattedBody
                                                               isSenderMessageAnEmote:isSenderMessageAnEmote
-                                                               replyFormattedMessage:finalFormattedTextMessage
-                                                                     stringLocalizer:stringLocalizer];
+                                                               replyFormattedMessage:finalFormattedTextMessage];
     }
 }
 
@@ -2436,7 +2465,6 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
  @param senderMessageFormattedBody The message body of the sender.
  @param isSenderMessageAnEmote Indicate if the sender message is an emote (/me).
  @param replyFormattedMessage The response for the sender message. HTML formatted string if any otherwise non formatted string as reply formatted body is mandatory.
- @param stringLocalizer string localizations used when building formatted body.
  
  @return reply message body.
  */
@@ -2444,7 +2472,6 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
                             senderMessageFormattedBody:(NSString*)senderMessageFormattedBody
                                 isSenderMessageAnEmote:(BOOL)isSenderMessageAnEmote
                                  replyFormattedMessage:(NSString*)replyFormattedMessage
-                                       stringLocalizer:(id<MXSendReplyEventStringLocalizerProtocol>)stringLocalizer
 {
     NSString *eventId = eventToReply.eventId;
     NSString *roomId = eventToReply.roomId;
@@ -2490,7 +2517,9 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
     [replyMessageFormattedBody appendString:@"<mx-reply><blockquote>"];
     
     // Add event link
-    [replyMessageFormattedBody appendFormat:@"<a href=\"%@\">%@</a> ", eventPermalink, stringLocalizer.messageToReplyToPrefix];
+    // The "In reply to" string is not meant to be localized from the sender side.
+    // This is how here we use the default string localizer to send the english version of it.
+    [replyMessageFormattedBody appendFormat:@"<a href=\"%@\">%@</a> ", eventPermalink, MXSendReplyEventDefaultStringLocalizer.new.messageToReplyToPrefix];
     
     if (isSenderMessageAnEmote)
     {
@@ -2516,12 +2545,17 @@ NSInteger const kMXRoomInvalidInviteSenderErrorCode = 9002;
 
 - (BOOL)canReplyToEvent:(MXEvent *)eventToReply
 {
-    if(eventToReply.eventType == MXEventTypePollStart)
+    if (eventToReply.eventType == MXEventTypePollStart)
     {
         return YES;
     }
     
-    if(eventToReply.eventType == MXEventTypeBeaconInfo)
+    if (eventToReply.eventType == MXEventTypePollEnd)
+    {
+        return YES;
+    }
+    
+    if (eventToReply.eventType == MXEventTypeBeaconInfo)
     {
         return YES;
     }