+++ MatrixSDK/MXRestClient.h
@@ -91,6 +91,11 @@ FOUNDATION_EXPORT NSString *const kMXAccountDataTypeRecentRoomsKey;
 FOUNDATION_EXPORT NSString *const kMXAccountDataLocalNotificationKeyPrefix;
 FOUNDATION_EXPORT NSString *const kMXAccountDataIsSilencedKey;
 
+/**
+ Threads list request parameters
+ */
+FOUNDATION_EXPORT NSString *const kMXThreadsListIncludeAllParameter;
+FOUNDATION_EXPORT NSString *const kMXThreadsListIncludeParticipatedParameter;
 
 /**
  MXRestClient error domain
@@ -618,6 +623,19 @@ NS_REFINED_FOR_SWIFT;
                            success:(void (^)(void))success
                            failure:(void (^)(NSError *error))failure NS_REFINED_FOR_SWIFT;
 
+/**
+ Delete an account_data event for the client.
+
+ @param type The event type of the account_data to delete (@see kMXAccountDataType* strings)
+ Custom types should be namespaced to avoid clashes.
+ @param success A block object called when the operation succeeds.
+ @param failure A block object called when the operation fails.
+
+ @return a MXHTTPOperation instance.
+ */
+- (MXHTTPOperation*)deleteAccountDataWithType:(NSString*)type
+                                      success:(void (^)(void))success
+                                      failure:(void (^)(NSError *error))failure;
 
 #pragma mark - Filtering
 /**
@@ -1553,6 +1571,34 @@ NS_REFINED_FOR_SWIFT;
                         failure:(void (^)(NSError *error))failure NS_REFINED_FOR_SWIFT;
 
 /**
+ Redact an event and all related events in a room.
+ 
+ You can check whether a homeserver supports the redaction with relations via
+ `supportsRedactionWithRelations`.
+ 
+ @param eventId the id of the redacted event.
+ @param roomId the id of the room.
+ @param reason the redaction reason (optional).
+ @param txnId the transaction id to use. If nil, one will be generated.
+ @param relations the list of relation types (optional). If nil or empty, related events will not be redacted.
+ @param withRelationsIsStable whether the feature to redact related event is stable.
+
+ @param success A block object called when the operation succeeds.
+ @param failure A block object called when the operation fails.
+
+ @return a MXHTTPOperation instance.
+ */
+- (MXHTTPOperation*)redactEvent:(NSString*)eventId
+                         inRoom:(NSString*)roomId
+                         reason:(NSString*)reason
+                          txnId:(NSString*)txnId
+                  withRelations:(NSArray<NSString *>*)relations
+          withRelationsIsStable:(BOOL)withRelationsIsStable
+                        success:(void (^)(void))success
+                        failure:(void (^)(NSError *error))failure;
+
+
+/**
  Report an event.
 
  @param eventId the id of the event event.
@@ -1700,6 +1746,23 @@ NS_REFINED_FOR_SWIFT;
                                    to:(NSString*)roomVersion
                               success:(void (^)(NSString *replacementRoomId))success
                               failure:(void (^)(NSError *error))failure NS_REFINED_FOR_SWIFT;
+
+/**
+ List all the threads of a room.
+ 
+ @param roomId the id of the room.
+ @param include wether the response should include all threads (e.g. `kMXThreadsListIncludeAllParameter`) or only threads participated by the user (e.g. `kMXThreadsListIncludeParticipatedParameter`)
+ @param from the token to pass for doing pagination from a previous response.
+ @param success A block object called when the operation succeeds. It provides the list of root events of the threads and, optionally, the next batch token.
+ @param failure A block object called when the operation fails.
+ 
+ @return a MXHTTPOperation instance.
+ */
+- (MXHTTPOperation*)threadsInRoomWithId:(NSString*)roomId
+                                include:(NSString *)include
+                                   from:(nullable NSString*)from
+                                success:(void (^)(MXAggregationPaginatedResponse *response))success
+                                failure:(void (^)(NSError *error))failure NS_REFINED_FOR_SWIFT;
 
 #pragma mark - Room tags operations
 /**